refactor(date-utils): unifica date e date-utils
também refatora `is_holiday` e `convert_date_to_text`
